Separate client creation into separate module.
