+++ src/wallabag.Data/ViewModels/LoginPageViewModel.cs
@@ -5,7 +5,6 @@ using SQLite.Net;
 using System;
 using System.Collections.Generic;
 using System.Net.Http;
-using System.Text;
 using System.Threading.Tasks;
 using wallabag.Api;
 using wallabag.Api.Models;
@@ -23,6 +22,7 @@ namespace wallabag.Data.ViewModels
         private readonly INavigationService _navigationService;
         private readonly IPlatformSpecific _device;
         private readonly IWallabagClient _client;
+        private readonly IApiClientCreationService _apiClientService;
         private readonly SQLiteConnection _database;
 
         private bool _restoredFromPageState = false;
@@ -57,12 +57,14 @@ namespace wallabag.Data.ViewModels
             INavigationService navigation,
             IPlatformSpecific device,
             IWallabagClient client,
+            IApiClientCreationService apiService,
             SQLiteConnection database)
         {
             _loggingService = logging;
             _navigationService = navigation;
             _device = device;
             _client = client;
+            _apiClientService = apiService;
             _database = database;
 
             _loggingService.WriteLine("Creating new instance of LoginPageViewModel.");
@@ -210,7 +212,7 @@ namespace wallabag.Data.ViewModels
                 _client.InstanceUri = new Uri(Url);
                 _loggingService.WriteLine($"Instance URI of the client: {_client.InstanceUri}");
 
-                bool clientCreationIsSuccessful = await CreateApiClientAsync();
+                bool clientCreationIsSuccessful = (await _apiClientService.CreateClientAsync(Url, Username, Password)).Success;
                 _loggingService.WriteLineIf(clientCreationIsSuccessful, "Client creation is successful.");
 
                 if (clientCreationIsSuccessful == false &&
@@ -220,7 +222,7 @@ namespace wallabag.Data.ViewModels
                     Url = Url.Replace("https://", "http://");
                     _client.InstanceUri = new Uri(Url);
 
-                    if (await CreateApiClientAsync() == false)
+                    if ((await _apiClientService.CreateClientAsync(Url, Username, Password)).Success == false)
                     {
                         _loggingService.WriteLine("Client creation failed.");
                         return false;
@@ -368,194 +370,5 @@ namespace wallabag.Data.ViewModels
 
             return Task.FromResult(true);
         }
-
-        #region Client creation
-
-        HttpClient _http;
-
-        private const string m_LOGINSTARTSTRING = "<input type=\"hidden\" name=\"_csrf_token\" value=\"";
-        private const string m_TOKENSTARTSTRING = "<input type=\"hidden\" id=\"client__token\" name=\"client[_token]\" value=\"";
-        private const string m_FINALTOKENSTARTSTRING = "<strong><pre>";
-        private const string m_FINALTOKENENDSTRING = "</pre></strong>";
-        private const string m_HTMLINPUTENDSTRING = "\" />";
-
-        public async Task<bool> CreateApiClientAsync()
-        {
-            _loggingService.WriteLine("Creating a new client...");
-            ProgressDescription = _device.GetLocalizedResource("CreatingClientMessage");
-
-            string token = string.Empty;
-            bool useNewApi = false;
-            int step = 1;
-            HttpResponseMessage message = null;
-
-            try
-            {
-                _http = new HttpClient();
-                var instanceUri = new Uri(Url);
-
-                _loggingService.WriteLine("Logging in to get a cookie... (mmh, cookies...)");
-                _loggingService.WriteLine($"URI: {instanceUri.Append("/login_check")}");
-
-                // Step 1: Login to get a cookie.
-                var loginContent = new StringContent($"_username={System.Net.WebUtility.UrlEncode(Username)}&_password={System.Net.WebUtility.UrlEncode(Password)}&_csrf_token={await GetCsrfTokenAsync()}", Encoding.UTF8, "application/x-www-form-urlencoded");
-                var loginResponse = await _http.PostAsync(instanceUri.Append("/login_check"), loginContent);
-
-                if (!loginResponse.IsSuccessStatusCode)
-                {
-                    _loggingService.WriteLine($"Failed. Resulted content: {await loginResponse.Content.ReadAsStringAsync()}", LoggingCategory.Warning);
-                    return false;
-                }
-
-                // Step 2: Get the client token
-                _loggingService.WriteLine("Get the client token...");
-                step++;
-                var clientCreateUri = instanceUri.Append("/developer/client/create");
-                token = await GetStringFromHtmlSequenceAsync(clientCreateUri, m_TOKENSTARTSTRING, m_HTMLINPUTENDSTRING);
-
-                _loggingService.WriteLine($"URI: {clientCreateUri}");
-                _loggingService.WriteLine($"Token: {token}");
-
-                // Step 3: Create the new client
-                _loggingService.WriteLine("Creating the new client...");
-                step++;
-                string stringContent = string.Empty;
-                useNewApi = (await _client.GetVersionAsync()).Minor > 0;
-
-                _loggingService.WriteLine($"Use new API: {useNewApi}");
-
-                stringContent = $"client[redirect_uris]={GetRedirectUri(useNewApi)}&client[save]=&client[_token]={token}";
-
-                if (useNewApi)
-                    stringContent = $"client[name]={_device.DeviceName}&" + stringContent;
-
-                _loggingService.WriteLine($"Content: {stringContent}");
-
-                var addContent = new StringContent(stringContent, Encoding.UTF8, "application/x-www-form-urlencoded");
-                var addResponse = _http.PostAsync(clientCreateUri, addContent);
-
-                message = await addResponse;
-
-                if (!message.IsSuccessStatusCode)
-                {
-                    _loggingService.WriteLine($"Failed. Resulted content: {await message.Content.ReadAsStringAsync()}", LoggingCategory.Warning);
-                    return false;
-                }
-
-                string content = await message.Content.ReadAsStringAsync();
-                _loggingService.WriteLine($"Parsing the resulted string: {content}");
-
-                var result = ParseResult(content, useNewApi);
-                if (result != null)
-                {
-                    _loggingService.WriteLine("Success!");
-                    ClientId = result.Id;
-                    ClientSecret = result.Secret;
-
-                    _http.Dispose();
-                }
-                else
-                    return false;
-
-                return true;
-            }
-            catch (Exception e)
-            {
-                _loggingService.TrackException(e);
-                return false;
-            }
-        }
-
-        public Uri GetRedirectUri(bool useNewApi) => useNewApi ? default(Uri) : new Uri(Url).Append(_device.DeviceName);
-        private Task<string> GetCsrfTokenAsync() => GetStringFromHtmlSequenceAsync(new Uri(Url).Append("/login"), m_LOGINSTARTSTRING, m_HTMLINPUTENDSTRING);
-
-        private async Task<string> GetStringFromHtmlSequenceAsync(Uri uri, string startString, string endString)
-        {
-            _loggingService.WriteLine("Trying to get a string from HTML code.");
-            _loggingService.WriteLine($"URI: {uri}");
-            _loggingService.WriteLine($"Start string: {startString}");
-            _loggingService.WriteLine($"End string: {endString}");
-
-            string html = await (await _http.GetAsync(uri)).Content.ReadAsStringAsync();
-            _loggingService.WriteLine($"HTML to parse: {html}");
-
-            int startIndex = html.IndexOf(startString) + startString.Length;
-            int endIndex = html.IndexOf(endString, startIndex);
-
-            _loggingService.WriteLine($"Start index: {startIndex}");
-            _loggingService.WriteLine($"End index: {endIndex}");
-
-            string result = html.Substring(startIndex, endIndex - startIndex);
-            _loggingService.WriteLine($"Result: {result}");
-
-            return result;
-        }
-
-        private ClientResultData ParseResult(string html, bool useNewApi = false)
-        {
-            _loggingService.WriteLine("Trying to parse the resulted client credentials...");
-            _loggingService.WriteLine($"Use new API: {useNewApi}");
-            _loggingService.WriteLine($"HTML to parse: {html}");
-
-            try
-            {
-                var results = new List<string>();
-                int resultCount = useNewApi ? 2 : 1;
-
-                int lastIndex = 0;
-
-                do
-                {
-                    int start = html.IndexOf(m_FINALTOKENSTARTSTRING, lastIndex) + m_FINALTOKENSTARTSTRING.Length;
-                    lastIndex = html.IndexOf(m_FINALTOKENENDSTRING, start);
-
-                    _loggingService.WriteLine($"Start index: {start}");
-                    _loggingService.WriteLine($"Last index: {lastIndex}");
-
-                    string result = html.Substring(start, lastIndex - start);
-
-                    _loggingService.WriteLine($"Result: {result}");
-
-                    results.Add(result);
-
-                    _loggingService.WriteLine($"Number of results: {results.Count}");
-
-                } while (results.Count <= resultCount);
-
-                var finalResult = default(ClientResultData);
-                if (useNewApi)
-                    finalResult = new ClientResultData()
-                    {
-                        Name = results[0],
-                        Id = results[1],
-                        Secret = results[2]
-                    };
-                else
-                    finalResult = new ClientResultData()
-                    {
-                        Id = results[0],
-                        Secret = results[1],
-                        Name = string.Empty
-                    };
-
-                _loggingService.WriteObject(finalResult);
-                return finalResult;
-            }
-            catch (Exception e)
-            {
-                _loggingService.TrackException(e);
-                Messenger.Default.Send(new NotificationMessage(_device.GetLocalizedResource("SomethingWentWrongMessage")));
-                return null;
-            }
-        }
-
-        #endregion
-    }
-
-    public class ClientResultData
-    {
-        public string Id { get; set; }
-        public string Secret { get; set; }
-        public string Name { get; set; }
     }
 }
